// app/settings/components/workspace-reconcile/WorkspaceReconcileCard.module.css.d.ts
declare const styles: Readonly<{
  reconcileBody: string;
  toolbar: string;
  actionList: string;
  checkedAt: string;
  readonly: string;
  status: string;
  informational: string;
  error: string;
  inSync: string;
  reconcileTable: string;
  stateChip: string;
  missing: string;
  renamed: string;
  unmanaged: string;
  names: string;
}>;

export default styles;